Accept/requirements.txt
asgiref==3.6.0
Django==4.2.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
Pillow==9.5.0
psycopg==3.1.9
sqlparse==0.4.4
typing_extensions==4.5.0
